feat(ingestion): measure upsert con detección de cambios, control de revisión e ingesta tolerante
Upsert por fila vía payload_hash (insert / skip / update-in-place conservando el
id), guard de precedencia revisión/iteración (gana la mayor; archivo viejo no pisa
datos nuevos), pre-check cross-familia que rechaza antes de escribir, y tres capas
de error tolerantes (parseo, validación, binary-split de BD con commit parcial real
vía REQUIRES_NEW). Línea de evento por archivo reconciliable
(total = persisted + updated + skipped + parseDefects + validationDefects + quarantined).

Almacenamiento dimensionado al mínimo: payload_hash BINARY(8), columnas right-sized,
sin auditoría redundante, particionado anual por fecha + job de mantenimiento en
ghost-flows. Resolución de clientes en una sola consulta (sin N+1); ids TSID
generados en app para habilitar JDBC batching. P1 redondeado a techo.

Validado E2E contra MySQL real (12 escenarios, ver docs/upsert-medidas-escenarios-e2e.md)
+ 129 tests unitarios; reactor build OK.

+++ c4e-ingestion-service/src/main/java/com/com4energy/processor/service/measure/MeasureFileParserService.java
@@ -90,23 +90,23 @@ public class MeasureFileParserService {
             String cups = cursor.nextRaw();
             int tipoMedida = cursor.nextInt(metadata.originalFilename(), lineNumber);
             LocalDateTime timestamp = cursor.nextTimestamp(metadata.originalFilename(), lineNumber);
-            float banderaInvVer = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float actent = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float qactent = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float actsal = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float qactsal = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float rQ1 = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float qrQ1 = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float rQ2 = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float qrQ2 = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float rQ3 = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float qrQ3 = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float rQ4 = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float qrQ4 = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float medres1 = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float qmedres1 = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float medres2 = cursor.nextFloat(metadata.originalFilename(), lineNumber);
-            float qmedres2 = cursor.nextFloat(metadata.originalFilename(), lineNumber);
+            double banderaInvVer = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double actent = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double qactent = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double actsal = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double qactsal = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double rQ1 = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double qrQ1 = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double rQ2 = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double qrQ2 = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double rQ3 = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double qrQ3 = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double rQ4 = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double qrQ4 = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double medres1 = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double qmedres1 = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double medres2 = cursor.nextDouble(metadata.originalFilename(), lineNumber);
+            double qmedres2 = cursor.nextDouble(metadata.originalFilename(), lineNumber);
             int metodObt = cursor.nextInt(metadata.originalFilename(), lineNumber);
             int temporal = cursor.nextInt(metadata.originalFilename(), lineNumber);
 
@@ -282,10 +282,10 @@ public class MeasureFileParserService {
             }
         }
 
-        private float nextFloat(String fileName, int lineNumber) {
+        private double nextDouble(String fileName, int lineNumber) {
             String value = nextRaw();
             try {
-                return Float.parseFloat(value);
+                return Double.parseDouble(value);
             } catch (NumberFormatException ex) {
                 throw conversionError(value, fileName, lineNumber);
             }
